refactor(firestore): consolidate atomic write policy
